// src/findora/apis/types.ts
export interface OwnedSidsDataResult extends NetworkAxiosDataResult {
  response?: number[];
}

export type OwnedMemoResponse = {
  blind_share: string;
  lock: {
    ciphertext: string;
    ephemeral_public_key: string;
  };
};

export interface OwnerMemoDataResult extends NetworkAxiosDataResult {
  response?: OwnedMemoResponse;
}

export type MTleafNode = {
  siblings1: string;
  siblings2: string;
  is_left_child: number;
  is_right_child: number;
};

export type MTLeafInfoResponse = {
  uid: string;
  root_version: string;
  root: string;
  path: {
    nodes: MTleafNode[];
  };
};

export interface MTLeafInfoDataResult extends NetworkAxiosDataResult {
  response?: MTLeafInfoResponse;
}

export interface LedgerUtxo {
  id?: number | null | undefined;
  record: any;
}

export type UtxoResponse = {
  utxo: LedgerUtxo;
  authenticated_txn?: string;
  finalized_txn?: any;
  txn_inclusion_proof?: any;
  state_commitment_data?: any;
  state_commitment?: string[];
};

export interface UtxoDataResult extends NetworkAxiosDataResult {
  response?: UtxoResponse;
}

export type OwnedResponseAbarItem = [string, FindoraWallet.OwnedAbar];

export type OwnedAbarsResponse = OwnedResponseAbarItem;

export interface OwnedAbarsDataResult extends NetworkAxiosDataResult {
  response?: OwnedAbarsResponse;
}

export interface CheckNullifierHashSpentDataResult extends NetworkAxiosDataResult {
  response?: boolean;
}

export type AssetTokenResponse = {
  properties: FindoraWallet.IPureAsset;
};

export interface AssetTokenDataResult extends NetworkAxiosDataResult {
  response?: AssetTokenResponse;
}

export type BlockDetailsResponse = {
  result: {
    block_id: {
      hash: string;
    };
    block: {
      header: {
        chain_id: string;
        height: string;
        time: string | undefined;
      };
      data: {
        txs: null | any[];
      };
    };
  };
};

export interface BlockDetailsDataResult extends NetworkAxiosDataResult {
  response?: BlockDetailsResponse;
}

export interface TxResult {
  code: number;
  hash: string;
  time: string | undefined;
  data: null | any[];
  log?: string;
  info?: string;
  gasWanted?: string;
  gasUsed?: string;
}

export interface AbciNoce {
  code: number;
  log: string;
  info: string;
  index: string;
  key: any;
  value: string;
  proof: any;
  height: string;
  codespace: string;
}

export interface SubmitEvmTx {
  height: string;
  hash: string;
  deliver_tx: { log: string; data: string; code: string };
  check_tx: {
    log: string;
    data: string;
    code: string;
  };
}

export interface TxInfo {
  code: number;
  data: null | any[];
  hash: string;
  time: string | undefined;
  height: number;
  index: number;
  tx_result: TxResult;
  tx: string;
}
export interface TxListResponseResult {
  txs: null | TxInfo[];
  total_count: number;
}

export type TxListResponse = {
  result: TxListResponseResult;
};

export interface TxListDataResult extends NetworkAxiosDataResult {
  response?: TxListResponse;
}

export type TxDetailsResponse = {
  result: {
    tx: string;
  };
};

export interface TxDetailsDataResult extends NetworkAxiosDataResult {
  response?: TxDetailsResponse;
}

export type HashSwapResponse = {
  result: TxListResponseResult;
};
// ...
export interface HashSwapDataResult extends NetworkAxiosDataResult {
  response?: HashSwapResponse;
}

export type StateCommitmenResponse = [number[], number, string];

export interface StateCommitmentDataResult extends NetworkAxiosDataResult {
  response?: StateCommitmenResponse;
}

export interface ValidatorListDataResult extends NetworkAxiosDataResult {
  response?: ValidatorListResponse;
}

export interface DelegateInfoDataResult extends NetworkAxiosDataResult {
  response?: DelegateInfoResponse;
}

export type TransactionData = string;

export interface ParsedTransactionData {}

export interface SubmitTransactionDataResult extends NetworkAxiosDataResult {
  response?: string;
}
// ...
export type AbciNoceResponse = {
  result: { response: AbciNoce };
};

export interface AbciNoceResult extends NetworkAxiosDataResult {
  response?: AbciNoceResponse;
}

export interface AbciInfoResult extends NetworkAxiosDataResult {
  response?: AbciNoceResponse;
}